requirements.txt
packaging==24.2
pillow==11.0.0
PyMuPDF==1.25.1
pytesseract==0.3.13